add x_space.h  add includes
