+++ src/apps/molresponse/TDDFT.cc
@@ -25,6 +25,7 @@
 #include "Plot_VTK.h"
 #include "TDHF_Basic_Operators2.h"
 #include "molresponse/response_functions.h"
+#include "molresponse/x_space.h"
 #include "molresponse/density.h"
 #include "molresponse/global_functions.h"
 #include "molresponse/property.h"
