+++ src/apps/molresponse/TDDFT.h
@@ -62,7 +62,8 @@
 #include "../chem/molecule.h"
 #include "../chem/xcfunctional.h"
 #include "GroundParameters.h"
-#include "response_functions.h"
+#include "molresponse/response_functions.h"
+#include "molresponse/x_space.h"
 #include "ResponseParameters.h"
 #include "ResponsePotential.h"
 #include "TDHF_Basic_Operators2.h"
