+++ src/apps/molresponse/iterate_gamma.cc
@@ -15,6 +15,7 @@
 #include "TDDFT.h"
 #include "TDHF_Basic_Operators2.h"
 #include "molresponse/response_functions.h"
+#include "molresponse/x_space.h"
 #include "molresponse/density.h"
 #include "molresponse/global_functions.h"
 #include "molresponse/property.h"
