+++ src/apps/molresponse/iterate_xy.cc
@@ -14,6 +14,7 @@
 #include "TDDFT.h"
 #include "TDHF_Basic_Operators2.h"
 #include "molresponse/response_functions.h"
+#include "molresponse/x_space.h"
 #include "molresponse/density.h"
 #include "molresponse/global_functions.h"
 #include "molresponse/property.h"
